refactor(epicenter): rename y-doc-sync to update-ytext
The old name described the mechanism (doc sync) rather than the intent
(updating Y.Text from a string using minimal diffs). The new name makes
the utility discoverable for its actual use case: CRDT-preserving text
updates via character-level diffing.
